refactor(settings): extract BehaviorsSettings panel (#1600, part 2)
Second slice of the SettingsDialog panel split. The behaviors panel is fully
self-contained (per-change persistence, no Done-batch, no parent props), so it
lifts out cleanly:

- New BehaviorsSettings.svelte owns the sidebar / breadcrumbs / conversations
  toggles + the confirmation-dialog suppression list (state, patch*, confirm
  store, derived rows, template, styles).
- makePatch moved from SettingsDialog into a shared util (src/renderer/lib/
  make-patch.ts) so both it and the new panel use it; SettingsDialog keeps
  patchRefactor (notes panel).
- SettingsDialog renders <BehaviorsSettings /> and sheds the moved state +
  imports (−110 lines).

Pure extraction — behaviour identical (moved verbatim). pnpm lint clean; full
suite green (5424).
